refactor: widgets 제거
