+++ frontend/techpick/src/components/DirectoryNode/DirectoryNode.css.ts
