+++ frontend/techpick/src/components/FeaturedSection/FeaturedSection.css.ts
