+++ frontend/techpick/src/components/FeaturedSection/FeaturedSection.tsx
