+++ frontend/techpick/src/components/TagPicker/TagAutocompleteDialog/useCalculateCommandListHeight.ts
