+++ frontend/techpick/src/components/index.ts
@@ -12,3 +12,10 @@ export { ShowDeleteTagDialogButton } from './ShowDeleteTagDialogButton';
 export { ToggleThemeButton } from './ToggleThemeButton';
 export { Folder } from './nodeManagement/ui/Folder';
 export { Pick } from './nodeManagement/ui/Pick';
+export { EditorContextMenu } from './ContextMenu/EditorContextMenu';
+export { TreeContextMenu } from './ContextMenu/TreeContextMenu';
+export { DirectoryNode } from './DirectoryNode/DirectoryNode';
+export { DirectoryTreeSection } from './DirectoryTreeSection/DirectoryTreeSection';
+export { FeaturedSection } from './FeaturedSection/FeaturedSection';
+export { LinkEditorSection } from './LinkEditorSection/index';
+export { TagPicker } from './TagPicker';
