--- frontend/techpick/src/widgets/index.ts
@@ -1,3 +1,0 @@
-export { DirectoryTreeSection } from '@/widgets/DirectoryTreeSection/DirectoryTreeSection';
-export { LinkEditorSection } from '@/widgets/LinkEditorSection';
-export { FeaturedSection } from '@/widgets/FeaturedSection/FeaturedSection';
